The Pathscale compiler on the MSU HPC cluster has problems with log function using integer valse and the pow function using unsigned integers as exponents. Modified the code to accomidate this.

+++ source/cpu/cHardwareCPU.cc
@@ -2744,7 +2744,7 @@ bool cHardwareCPU::Inst_Sense(cAvidaContext& ctx)
  
   if ((last_num_resources != res_count.GetSize()))
   {
-      int max_label_length = (int) ceil(log(res_count.GetSize())/log(cHardwareCPU::NUM_NOPS));
+      int max_label_length = (int) ceil(log((float) res_count.GetSize())/log((float) cHardwareCPU::NUM_NOPS));
       last_num_resources = res_count.GetSize();
   }
